added constructor that takes a pointer to double array
operator+ and operator- now create a new heap-allocated double* array and uses that to construct the new matrix from the operation result, using the new matrix constructor that takes a pointer to double array. no longer creates a new vector and fills that out for the vector constructor.

gitignore update is random extension i just found, ignore it