LOAD_PATIENTS flag, Application::quit()
In application.cpp, you can now set LOAD_PATIENTS to 0 to disable loading patient data
or 1 to load them in an construct the vector of Patient objects

For debugging purposes.

Also added quit function to application class so that it can be called anywhere
for whatever reason. Use Application::getInstance()->quit()